fix screens, full end to end flow working
