+++ .eslintrc.js
@@ -2,9 +2,9 @@ module.exports = {
   extends: [
     'plugin:@typescript-eslint/recommended',
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
-    'plugin:prettier/recommended',
     'plugin:react/recommended',
     'prettier/@typescript-eslint',
+    'plugin:prettier/recommended',
     // 'plugin:jsx-a11y/recommended'
   ],
   parser: '@typescript-eslint/parser',
@@ -12,13 +12,13 @@ module.exports = {
     ecmaVersion: '2018',
     sourceType: 'module',
     project: './tsconfig.json',
-    tsconfigRootDir: __dirname
+    tsconfigRootDir: __dirname,
   },
   plugins: ['@typescript-eslint', 'react'],
   env: {
     browser: true,
     node: true,
-    es6: true
+    es6: true,
   },
   rules: {
     quotes: [2, 'single', { avoidEscape: true }],
@@ -33,17 +33,16 @@ module.exports = {
       {
         multiline: {
           delimiter: 'semi',
-          requireLast: true
+          requireLast: true,
         },
         singleline: {
           delimiter: 'semi',
-          requireLast: false
-        }
-      }
+          requireLast: false,
+        },
+      },
     ],
     '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
     '@typescript-eslint/array-type': 'error',
-    // '@typescript-eslint/semi': ['error', 'never'],
     '@typescript-eslint/no-misused-promises': 'error',
     '@typescript-eslint/no-floating-promises': 'error',
     '@typescript-eslint/require-await': 'error',
@@ -62,5 +61,5 @@ module.exports = {
     'react/jsx-uses-vars': [2],
     'react/jsx-key': [0],
     'react/prop-types': [0],
-  }
-}
+  },
+};
