+++ package.json
@@ -17,19 +17,20 @@
     "test:coverage": "jest --collect-coverage",
     "typecheck": "tsc --noEmit"
   },
+  "prettier": "@blockstack/prettier-config",
   "license": "MIT",
   "dependencies": {
     "@blockstack/connect": "^0.2.0-beta.1",
     "@blockstack/keychain": "^0.1.2",
     "@blockstack/prettier-config": "^0.0.5",
-    "@blockstack/ui": "^1.0.0-alpha.23",
+    "@blockstack/ui": "^1.0.0-alpha.24",
     "formik": "^2.0.3",
     "mdi-react": "^6.3.0",
     "react": "^16.12.0",
     "react-chrome-redux": "^2.0.0-alpha.5",
     "react-dom": "^16.12.0",
     "react-redux": "^7.1.1",
-    "redux": "4.0.4",
+    "redux": "4.0.5",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.3.0",
     "styled-components": "^5.0.0",
@@ -48,14 +49,14 @@
     "@babel/preset-typescript": "^7.6.0",
     "@babel/runtime": "^7.6.3",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.1.3",
-    "@types/chrome": "^0.0.86",
-    "@types/jest": "^24.0.15",
-    "@types/node": "^12.6.2",
+    "@types/chrome": "^0.0.91",
+    "@types/jest": "^24.9.0",
+    "@types/node": "^13.1.7",
     "@types/react": "^16.9.17",
     "@types/react-dom": "^16.9.4",
     "@types/react-redux": "^7.1.2",
     "@types/react-test-renderer": "^16.8.2",
-    "@types/styled-components": "4.1.8",
+    "@types/styled-components": "4.4.2",
     "@types/webpack": "^4.4.35",
     "@typescript-eslint/eslint-plugin": "^2.6.0",
     "@typescript-eslint/parser": "^2.6.0",
@@ -63,12 +64,12 @@
     "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^5.0.3",
-    "cross-env": "^5.2.0",
+    "cross-env": "^6.0.3",
     "eslint": "^6.3.0",
     "eslint-config-prettier": "^6.2.0",
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react": "^7.14.3",
-    "fork-ts-checker-webpack-plugin": "^1.5.0",
+    "fork-ts-checker-webpack-plugin": "^3.1.1",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.8.0",
     "prettier": "^1.19.1",
@@ -77,7 +78,7 @@
     "ts-jest": "^24.0.2",
     "ts-loader": "^6.0.4",
     "tslint": "^5.18.0",
-    "typescript": "3.7.2",
+    "typescript": "3.7.4",
     "webpack": "^4.41.2",
     "webpack-chrome-extension-reloader": "^1.3.0",
     "webpack-cli": "^3.3.5",
