+++ src/ts/components/sign-up/collapse/index.tsx
@@ -67,7 +67,7 @@ const Collapse: React.FC<CollapseProps> = ({ data, ...rest }) => {
   const [open, setOpen] = React.useState<number | null>(null);
   const handleOpen = (key: number) => (key === open ? setOpen(null) : setOpen(key));
   return (
-    <Box fontSize="14px" {...rest}>
+    <Box px={5} fontSize="14px" {...rest}>
       {/*
           It's important to include the rest of the props
           because in certain cases we want to add/adjust spacing,
