+++ src/ts/components/sign-up/modal/index.tsx
@@ -23,10 +23,7 @@ interface HeaderTitleProps {
   hideIcon?: boolean;
 }
 
-const HeaderTitle: React.FC<HeaderTitleProps> = ({
-  hideIcon = false,
-  title,
-}) => (
+const HeaderTitle: React.FC<HeaderTitleProps> = ({ hideIcon = false, title }) => (
   <Flex align="center">
     {hideIcon ? null : <Logo mr={2} />}
     <Text fontWeight="bold" fontSize={'12px'}>
@@ -36,12 +33,7 @@ const HeaderTitle: React.FC<HeaderTitleProps> = ({
 );
 
 const HeaderCloseButton: React.FC<BoxProps> = ({ onClick }) => (
-  <PseudoBox
-    color="ink.300"
-    opacity={0.5}
-    _hover={{ opacity: 1, cursor: 'pointer' }}
-    onClick={onClick}
-  >
+  <PseudoBox color="ink.300" opacity={0.5} _hover={{ opacity: 1, cursor: 'pointer' }} onClick={onClick}>
     <CloseIcon size={20} />
   </PseudoBox>
 );
@@ -65,14 +57,7 @@ interface IModalHeader {
   hideIcon?: boolean;
 }
 
-const ModalHeader = ({
-  appIcon,
-  close,
-  title,
-  hideIcon,
-  appName,
-  ...rest
-}: IModalHeader) => {
+const ModalHeader = ({ appIcon, close, title, hideIcon, appName, ...rest }: IModalHeader) => {
   return (
     <Flex
       p={[4, 5]}
@@ -115,15 +100,7 @@ interface ModalProps {
   close: () => void;
 }
 
-const Modal: React.FC<ModalProps> = ({
-  footer = null,
-  appIcon,
-  title,
-  hideIcon = false,
-  close,
-  appName,
-  children,
-}) => {
+const Modal: React.FC<ModalProps> = ({ footer = null, appIcon, title, hideIcon = false, close, appName, children }) => {
   return (
     <>
       <Flex
@@ -146,20 +123,8 @@ const Modal: React.FC<ModalProps> = ({
           borderRadius={['unset', '6px']}
           boxShadow="high"
         >
-          <ModalHeader
-            hideIcon={hideIcon}
-            close={close}
-            appIcon={appIcon}
-            appName={appName}
-            title={title}
-          />
-          <Flex
-            width="100%"
-            p={[5, 8]}
-            overflowY="auto"
-            flexGrow={1}
-            position="relative"
-          >
+          <ModalHeader hideIcon={hideIcon} close={close} appIcon={appIcon} appName={appName} title={title} />
+          <Flex width="100%" p={[5, 8]} overflowY="auto" flexGrow={1} position="relative">
             <ModalContent>{children}</ModalContent>
           </Flex>
           {footer ? footer : null}
